+++ 5333/src/main/include/RobotMap.h
@@ -4,8 +4,8 @@
 #include <frc/Spark.h>
 #include <frc/PowerDistributionPanel.h>
 
-#include "CurtinCtre.h"
-#include "controllers/CurtinControllers.h"
+#include "WMLCtre.h"
+#include "controllers/Controllers.h"
 #include "Gearbox.h"
 #include "actuators/BinaryServo.h"
 #include "actuators/Compressor.h"
