+++ 5333/src/main/include/SideHatch.h
@@ -6,7 +6,7 @@
 #include "strategy/StrategySystem.h"
 // #include "devices/DeployableDevice.h"
 #include "actuators/BinaryActuator.h"
-// #include "controllers/CurtinControllers.h"
+// #include "controllers/Controllers.h"
 // #include "Toggle.h"
 
 // using SideHatchState = wml::devices::DeployableDeviceState;
