+++ wml/src/main/cpp/controllers/Controllers.cpp
@@ -1,4 +1,4 @@
-#include "controllers/CurtinControllers.h"
+#include "controllers/Controllers.h"
 
 using namespace wml::controllers::inputs;
