+++ wml/src/simulation/include/simulation/ctre_sim.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "CurtinCtre.h"
+#include "WMLCtre.h"
 
 #include <map>
